Aula 33

+++ src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,9 @@ import {MatDialog} from "@angular/material/dialog";
 import {ErrorDialogComponent} from "../../../shared/components/error-dialog/error-dialog.component"
 import {ActivatedRoute, Router} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {
+  ConfirmationDialogComponent
+} from "../../../shared/components/confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-courses',
@@ -28,7 +31,7 @@ export class CoursesComponent {
   ngOnInit(): void {
   }
 
-  refresh(){
+  refresh() {
     this.courses$ = this.coursesService.list()
       .pipe(catchError(error => {
         this.onError('Erro ao carregar a lista.')
@@ -51,20 +54,29 @@ export class CoursesComponent {
   }
 
   onDelete(course: Course) {
-    this.coursesService.delete(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso deletado com sucesso!', '',
-          {
-            duration: 5000,
-            verticalPosition: "top",
-            horizontalPosition: "center"
-          })
-      },
-      error => {
-        this.onError('Erro ao deletar curso!');
-        return of([]);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.coursesService.delete(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso deletado com sucesso!', '',
+              {
+                duration: 5000,
+                verticalPosition: "top",
+                horizontalPosition: "center"
+              })
+          },
+          error => {
+            this.onError('Erro ao deletar curso!');
+            return of([]);
+          }
+        );
       }
-    );
+    });
+
   }
 }
